+++ src/lib/components/ExplorerLink.tsx
@@ -31,7 +31,6 @@ interface ExplorerLinkProps extends FlexProps {
   showCopyOnHover?: boolean;
   isReadOnly?: boolean;
   textFormat?: "truncate" | "ellipsis" | "normal";
-  maxWidth?: string;
   textVariant?: TextProps["variant"];
   ampCopierSection?: string;
   openNewTab?: boolean;
@@ -106,7 +105,6 @@ const LinkRender = ({
   hrefLink,
   textValue,
   isEllipsis,
-  maxWidth,
   textVariant,
   openNewTab,
 }: {
@@ -115,7 +113,6 @@ const LinkRender = ({
   hrefLink: string;
   textValue: string;
   isEllipsis: boolean;
-  maxWidth: ExplorerLinkProps["maxWidth"];
   textVariant: TextProps["variant"];
   openNewTab: Option<boolean>;
 }) => {
@@ -124,10 +121,7 @@ const LinkRender = ({
     <Text
       variant={textVariant}
       color="secondary.main"
-      transition="all 0.25s ease-in-out"
-      _hover={{ color: "secondary.light" }}
       className={isEllipsis ? "ellipsis" : undefined}
-      maxW={maxWidth}
       pointerEvents={hrefLink ? "auto" : "none"}
       wordBreak={{ base: "break-all", md: "inherit" }}
     >
@@ -136,7 +130,15 @@ const LinkRender = ({
   );
 
   return isInternal && !openNewTab ? (
-    <AppLink href={hrefLink} passHref onClick={(e) => e.stopPropagation()}>
+    <AppLink
+      href={hrefLink}
+      passHref
+      onClick={(e) => e.stopPropagation()}
+      style={{
+        whiteSpace: "nowrap",
+        overflow: "hidden",
+      }}
+    >
       {textElement}
     </AppLink>
   ) : (
@@ -148,6 +150,10 @@ const LinkRender = ({
       onClick={(e) => {
         if (!isInternal) trackMintScan(type);
         e.stopPropagation();
+      }}
+      style={{
+        whiteSpace: "nowrap",
+        overflow: "hidden",
       }}
     >
       {textElement}
@@ -163,7 +169,6 @@ export const ExplorerLink = ({
   showCopyOnHover = false,
   isReadOnly = false,
   textFormat = "truncate",
-  maxWidth = "160px",
   textVariant = "body2",
   ampCopierSection,
   openNewTab,
@@ -211,7 +216,6 @@ export const ExplorerLink = ({
         hrefLink={link}
         textValue={textValue}
         isEllipsis={textFormat === "ellipsis"}
-        maxWidth={maxWidth}
         textVariant={textVariant}
         openNewTab={openNewTab}
       />