+++ src/lib/pages/proposal-details/components/vote-details/voting-period/table/Voter.tsx
@@ -29,7 +29,6 @@ export const Voter = ({ proposalVote }: VoterProps) => {
             />
           ) : undefined
         }
-        maxWidth="110px"
       />
     );